Move VS Code debug server files
