+++ README.md
@@ -170,7 +170,7 @@ For complete setup instructions and tool documentation, see [MCP_README.md](MCP_
 A source-level debugger for VS Code -- **LynxDebug** -- lives in its own repository:
 [BrianPeek/gearlynx-vscode](https://github.com/BrianPeek/gearlynx-vscode). It drives
 Gearlynx over the debug-monitor protocol (`--debug-monitor`). The wire contract is
-documented in [PROTOCOL.md](PROTOCOL.md).
+documented in [README_VSCODE_PROTOCOL.md](README_VSCODE_PROTOCOL.md).
 
 ### Agent Skills
