+++ platforms/shared/desktop/vscode/framebuffer_server.h
