+++ platforms/windows/Gearlynx.vcxproj
@@ -66,8 +66,8 @@
     <ClInclude Include="..\shared\desktop\backers.h" />
     <ClInclude Include="..\shared\desktop\config.h" />
     <ClInclude Include="..\shared\desktop\console_utils.h" />
-    <ClInclude Include="..\shared\desktop\debug_monitor_server.h" />
-    <ClInclude Include="..\shared\desktop\framebuffer_server.h" />
+    <ClInclude Include="..\shared\desktop\vscode\debug_monitor_server.h" />
+    <ClInclude Include="..\shared\desktop\vscode\framebuffer_server.h" />
     <ClInclude Include="..\shared\desktop\socket_types.h" />
     <ClInclude Include="..\shared\desktop\emu.h" />
     <ClInclude Include="..\shared\desktop\gui.h" />
@@ -150,8 +150,8 @@
     <ClCompile Include="..\shared\desktop\application_headless.cpp" />
     <ClCompile Include="..\shared\desktop\events.cpp" />
     <ClCompile Include="..\shared\desktop\config.cpp" />
-    <ClCompile Include="..\shared\desktop\debug_monitor_server.cpp" />
-    <ClCompile Include="..\shared\desktop\framebuffer_server.cpp" />
+    <ClCompile Include="..\shared\desktop\vscode\debug_monitor_server.cpp" />
+    <ClCompile Include="..\shared\desktop\vscode\framebuffer_server.cpp" />
     <ClCompile Include="..\shared\desktop\emu.cpp" />
     <ClCompile Include="..\shared\desktop\gui.cpp" />
     <ClCompile Include="..\shared\desktop\gui_actions.cpp" />
